Issues Solved. Low: 194 195 198 202 206 Medium: 200 201 199 204 205 Add contacts in booking (almost done, can save to db)

+++ freightbit-lib/src/main/java/com/sr/biz/freightbit/customer/entity/Items.java
@@ -20,9 +20,9 @@ public class Items implements java.io.Serializable {
     private Integer customerId;
     private String itemName;
     private String itemCode;
-    private Integer length;
-    private Integer width;
-    private Integer height;
+    private Double length;
+    private Double width;
+    private Double height;
     private Float srp;
     private Integer criticalQuality;
     private Float basePrice;
@@ -37,7 +37,7 @@ public class Items implements java.io.Serializable {
 
     }
 
-    public Items(Integer customerItemsId, Client client, Integer customerId, String itemName, String itemCode, Integer length, Integer width, Integer height, Float srp, Integer criticalQuality, Float basePrice, String note, String description, Date createdTimeStamp, String createdBy, Date modifiedTimeStamp, String modifiedBy) {
+    public Items(Integer customerItemsId, Client client, Integer customerId, String itemName, String itemCode, Double length, Double width, Double height, Float srp, Integer criticalQuality, Float basePrice, String note, String description, Date createdTimeStamp, String createdBy, Date modifiedTimeStamp, String modifiedBy) {
         this.customerItemsId = customerItemsId;
         this.client = client;
         this.customerId = customerId;
@@ -83,16 +83,16 @@ public class Items implements java.io.Serializable {
     public void setItemCode(String itemCode) { this.itemCode = itemCode; }
 
     @Column(name = "length")
-    public Integer getLength() { return length; }
-    public void setLength(Integer length) { this.length = length; }
+    public Double getLength() { return length; }
+    public void setLength(Double length) { this.length = length; }
 
     @Column(name = "width")
-    public Integer getWidth() { return width; }
-    public void setWidth(Integer width) { this.width = width; }
+    public Double getWidth() { return width; }
+    public void setWidth(Double width) { this.width = width; }
 
     @Column(name = "height")
-    public Integer getHeight() { return height; }
-    public void setHeight(Integer height) { this.height = height; }
+    public Double getHeight() { return height; }
+    public void setHeight(Double height) { this.height = height; }
 
     @Column(name = "srp")
     public Float getSrp() { return srp; }
